requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & LLM
google-genai>=0.3.0        # New SDK for Gemini 2.x
google-generativeai>=0.4.0 # Legacy/Stable SDK support
# openai-whisper removed - YouTube transcription no longer needed
lancedb>=0.4.0             # 向量資料庫 (替換 ChromaDB)
pyarrow>=14.0.0            # LanceDB 依賴
sentence-transformers>=3.0.0  # EmbeddingGemma-300m 模型載入

# PDF Processing
pdfplumber>=0.11.0         # PDF 文字抽取 (用於 Fact Memory 導入)

# Web & Data
crawl4ai>=0.4.246          # 網頁爬取
ddgs>=8.0.0                # DuckDuckGo 搜尋 (Fact Checker)
beautifulsoup4>=4.12.0     # HTML 解析
# yt-dlp removed - YouTube download no longer needed

# Markdown & Content
markdown>=3.5.0            # HTML 表格轉換
python-frontmatter>=1.0.0  # YAML frontmatter 解析

# Utilities & Images
Pillow>=10.0.0             # 圖片處理與 WebP 轉換 (Visual Director)

# Testing
pytest>=7.4.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
